image/requirements.txt
numpy>=1.21.0
nltk>=3.7
transformers>=4.25.0
tensorflow-cpu>=2.10.0
torch>=1.12.0
nltk